refactor: fold push-token bookkeeping into PendingChanges; delete the file
PushHistoryTokenStore.swift named a phantom type (no PushHistoryTokenStore
exists). Its three token accessors (single caller: PendingChanges.swift) and its
one model, PushHistoryTokenRecord, all belong to the pending-changes/push flow —
so the whole bookkeeping unit moves into PendingChanges.swift and the file is
deleted. The accessors become private (only PendingChanges uses them); the model
stays internal (SyncContainer references it cross-file for schema registration,
which is registration, not ownership). 'Push' stays — it's the correct verb for
the last-pushed bookmark. Record doc trimmed to its contract.

Behavior-neutral: SwiftSync 172 (+30), DemoCore 43/43 green.
