+++ SwiftSync/Sources/SwiftSync/PendingChanges.swift
@@ -261,4 +261,63 @@ extension SwiftSync {
                     """)
         }
     }
+
+    /// `withPendingChanges` writes `PushHistoryTokenRecord` to advance its bookmark once the upload is
+    /// acknowledged. `SyncContainer` registers that model automatically; a caller who builds their own
+    /// `ModelContainer` may not. Validate it's in the schema *before* the upload, so an acknowledged
+    /// server write is never stranded by a token write that throws afterward.
+    private static func requireOfflinePushBookkeeping(in context: ModelContext) throws {
+        let name = String(describing: PushHistoryTokenRecord.self)
+        guard context.container.schema.entities.contains(where: { $0.name == name }) else {
+            throw SyncError.schemaValidation(
+                reason: """
+                    Offline push needs SwiftSync's bookkeeping model (\(name)) in the context's schema, \
+                    but it is missing. Build the container with SyncContainer, which registers it \
+                    automatically, so an acknowledged upload is never lost to a failed token write.
+                    """)
+        }
+    }
+
+    private static func lastPushedHistoryToken(for model: any PersistentModel.Type, in context: ModelContext)
+        -> DefaultHistoryToken?
+    {
+        let typeName = String(reflecting: model)
+        var descriptor = FetchDescriptor<PushHistoryTokenRecord>(predicate: #Predicate { $0.modelTypeName == typeName })
+        descriptor.fetchLimit = 1
+        guard let record = try? context.fetch(descriptor).first else { return nil }
+        return try? JSONDecoder().decode(DefaultHistoryToken.self, from: record.tokenData)
+    }
+
+    private static func setLastPushedHistoryToken(
+        _ token: DefaultHistoryToken, for model: any PersistentModel.Type, in context: ModelContext
+    )
+        throws
+    {
+        let typeName = String(reflecting: model)
+        guard let data = try? JSONEncoder().encode(token) else { return }
+        var descriptor = FetchDescriptor<PushHistoryTokenRecord>(predicate: #Predicate { $0.modelTypeName == typeName })
+        descriptor.fetchLimit = 1
+        if let record = try context.fetch(descriptor).first {
+            record.tokenData = data
+        } else {
+            context.insert(PushHistoryTokenRecord(modelTypeName: typeName, tokenData: data))
+        }
+        try context.save()
+    }
+}
+
+/// SwiftSync's "how far have I pushed" bookmark, one row **per model type** (not per data row) — the only
+/// durable offline state SwiftSync keeps. The change log itself is SwiftData history; this records the
+/// last-pushed `DefaultHistoryToken` so the pull can tell un-pushed local edits from already-pushed ones.
+@Model
+final class PushHistoryTokenRecord {
+    @Attribute(.unique) var modelTypeName: String
+    /// A JSON-encoded `DefaultHistoryToken`. Stored as `Data` because the token is a `Codable` value,
+    /// not a `@Model`, so SwiftData can't persist it directly.
+    var tokenData: Data
+
+    init(modelTypeName: String, tokenData: Data) {
+        self.modelTypeName = modelTypeName
+        self.tokenData = tokenData
+    }
 }
